+++ packages/react/src/index.tsx
@@ -1,3 +1,6 @@
 'use client'
 export * from './components/Button';
 export * from './components/Box';
+export * from './components/Center';
+export * from './components/Square';
+export * from './components/Circle';